requirements.txt
ccxt>=4.0.0
openai>=1.0.0
pandas>=2.0.0
schedule>=1.2.0
python-dotenv>=1.0.0
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
flask>=3.0.0
flask-socketio>=5.3.0
websocket-client>=1.7.0
sqlalchemy>=2.0.0
alembic>=1.13.0
plotly>=5.18.0
dash>=2.15.0
cachetools>=5.3.0
apscheduler>=3.10.0